version 1.5.4

+++ src/Crayon.Api.Sdk/Resources/UserResource.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System.Net;
+using System.Threading.Tasks;
 using Crayon.Api.Sdk.Domain;
 using Crayon.Api.Sdk.Filtering;
 
@@ -27,12 +28,22 @@ namespace Crayon.Api.Sdk.Resources
 
         public CrayonApiClientResult<User> GetById(string token, string id)
         {
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                throw new ApiHttpException(HttpStatusCode.BadRequest, "Id is required");
+            }
+
             var uri = $"/api/v1/users/{id}/";
             return _client.Get<User>(token, uri);
         }
 
         public async Task<CrayonApiClientResult<User>> GetByIdAsync(string token, string id)
         {
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                throw new ApiHttpException(HttpStatusCode.BadRequest, "Id is required");
+            }
+
             var uri = $"/api/v1/users/{id}/";
             return await _client.GetAsync<User>(token, uri);
         }
@@ -49,28 +60,39 @@ namespace Crayon.Api.Sdk.Resources
             return await _client.GetAsync<User>(token, uri);
         }
 
-        public CrayonApiClientResult<User> Create(string token, User user)
+        public CrayonApiClientResult<User> Create(string token, UserUpsert user)
         {
             var uri = "/api/v1/users/";
             return _client.Post<User>(token, uri, user);
         }
-        public async Task<CrayonApiClientResult<User>> CreateAsync(string token, User user)
+
+        public async Task<CrayonApiClientResult<User>> CreateAsync(string token, UserUpsert user)
         {
             var uri = "/api/v1/users/";
             return await _client.PostAsync<User>(token, uri, user);
         }
 
-        public CrayonApiClientResult<User> Update(string token, User user)
+        public CrayonApiClientResult<User> Update(string token, UserUpsert user)
         {
             Guard.NotNull(user, nameof(user));
+
+            if (string.IsNullOrWhiteSpace(user.Id))
+            {
+                throw new ApiHttpException(HttpStatusCode.BadRequest, "Id is required");
+            }
 
             var uri = $"/api/v1/users/{user.Id}/";
             return _client.Put<User>(token, uri, user);
         }
 
-        public async Task<CrayonApiClientResult<User>> UpdateAsync(string token, User user)
+        public async Task<CrayonApiClientResult<User>> UpdateAsync(string token, UserUpsert user)
         {
             Guard.NotNull(user, nameof(user));
+
+            if (string.IsNullOrWhiteSpace(user.Id))
+            {
+                throw new ApiHttpException(HttpStatusCode.BadRequest, "Id is required");
+            }
 
             var uri = $"/api/v1/users/{user.Id}/";
             return await _client.PutAsync<User>(token, uri, user);
@@ -78,11 +100,22 @@ namespace Crayon.Api.Sdk.Resources
 
         public CrayonApiClientResult Delete(string token, string id)
         {
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                throw new ApiHttpException(HttpStatusCode.BadRequest, "Id is required");
+            }
+
             var uri = $"/api/v1/users/{id}/";
             return _client.Delete(token, uri);
         }
+
         public async Task<CrayonApiClientResult> DeleteAsync(string token, string id)
         {
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                throw new ApiHttpException(HttpStatusCode.BadRequest, "Id is required");
+            }
+
             var uri = $"/api/v1/users/{id}/";
             return await _client.DeleteAsync(token, uri);
         }
@@ -94,6 +127,11 @@ namespace Crayon.Api.Sdk.Resources
 
         public CrayonApiClientResult<bool> ChangePassword(string token, string userId, string oldPassword, string newPassword)
         {
+            if (string.IsNullOrWhiteSpace(userId))
+            {
+                throw new ApiHttpException(HttpStatusCode.BadRequest, "Id is required");
+            }
+
             var uri = $"/api/v1/users/{userId}/changepassword/";
             return _client.Put<bool>(token, uri, new UserChangePassword { UserId = userId, OldPassword = oldPassword, NewPassword = newPassword });
         }
@@ -105,6 +143,10 @@ namespace Crayon.Api.Sdk.Resources
 
         public async Task<CrayonApiClientResult<bool>> ChangePasswordAsync(string token, string userId, string oldPassword, string newPassword)
         {
+            if (string.IsNullOrWhiteSpace(userId))
+            {
+                throw new ApiHttpException(HttpStatusCode.BadRequest, "Id is required");
+            }
             var uri = $"/api/v1/users/{userId}/changepassword/";
             return await _client.PutAsync<bool>(token, uri, new UserChangePassword { UserId = userId, OldPassword = oldPassword, NewPassword = newPassword });
         }
